Added an 'alias' extension
Aliasing commands allows one to create an alternative or short-hand
replacement to a command.  For example:
$ alias quit exit
Creates an alias named 'quit' which is expanded into 'exit' when the
command line is interpreted by the CLI interpreter.  Currently the
command auto-completion extension does not auto-complete aliased
command names.

Added support for quatation marks as separators:
	This is needed when you want a group of strings (words) to be
	treated as a unit.  For example, in
	$ alias ll "ls -la"
	We would like 'ls -la' to be treated as an atomic unit.

Allow write access to the LineEditor buffer:
	Some components need to change the command line content after it
	has been submitted to the interpreter.  For example, in case of
	vairable expansion (replacing $VAR with <var content>) or the 'alias'
	extension.