Add support for initialization script execution
Just like ~/.bashrc in Linux, we can add an initialization script
to set aliases and other shell parameters.
Of course, this is an example of running a shell initialization script,
but the underlying mechanism does not limit us to this use-case only.
In other words, you can run shell scripts at any time you want.

This commit also lays the groundwork for a 'set' command, which alters
the internal configuration of the shell itself.